pyproject.toml
[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = [
  "E", "F", "I",   # pycodestyle/pyflakes/isort
  "B",             # bugbear
  "UP",            # pyupgrade
]
ignore = ["E501"]  # Black ya gestiona line-length

[tool.ruff.lint.isort]
known-first-party = ["aetherya"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
pretty = true
mypy_path = "src"
packages = ["aetherya"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["src/aetherya"]

[tool.coverage.report]
fail_under = 99
show_missing = true
skip_covered = true

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "black",
  "mypy",
  "pyyaml",
  "types-PyYAML",
  "pre-commit",
]

[project]
name = "aetherya-core"
version = "0.3.0"
